+++ src/ui/views/Most.tsx
@@ -6,6 +6,17 @@ import { DataTable, PlayerNameLabels } from "../components";
 import type { View } from "../../common/types";
 import { frivolitiesMenu } from "./Frivolities";
 
+export const getValue = (
+	obj: any,
+	key: View<"most">["extraCols"][number]["key"],
+) => {
+	return typeof key === "string"
+		? obj[key]
+		: key.length === 2
+		? obj[key[0]][key[1]]
+		: obj[key[0]][key[1]][key[2]];
+};
+
 const Most = ({
 	challengeNoRatings,
 	description,
@@ -57,12 +68,7 @@ const Most = ({
 				p.rank,
 				<PlayerNameLabels pid={p.pid}>{p.name}</PlayerNameLabels>,
 				...extraCols.map(x => {
-					const value =
-						typeof x.key === "string"
-							? p[x.key]
-							: x.key.length === 2
-							? p[x.key[0]][x.key[1]]
-							: p[x.key[0]][x.key[1]][x.key[2]];
+					const value = getValue(p, x.key);
 					if (x.colName === "Amount") {
 						return helpers.formatCurrency(value / 1000, "M");
 					}