+++ CHANGELOG.md
@@ -2,7 +2,7 @@ This is a user-focused changelog. The entries here are things that affect gamepl
 
 Entries in bold are ones important enough that I would have put them in [the old changelog](./CHANGELOG-old.md).
 
-* DEV: New frivolity: Best Non-Playoff Teams.
+* DEV: New frivolities: Best Non-Playoff Teams, Worst Playoff Teams.
 
 * 2020-08-11: New frivolity: Draft Class Rankings. [[1]](https://old.reddit.com/r/BasketballGM/comments/i7kyx2/tools_frivolities_draft_class_rankings_new_in/)